add parser to document handler, so each file parsed on read or change
