+++ internal/lsp/completion.go
@@ -12,7 +12,7 @@ import (
 func textCompletion(context *glsp.Context, params *protocol.CompletionParams) (any, error) {
 	// executor := utils.CommandExecutor{}
 	uri := string(params.TextDocument.URI)
-	text := documents.Read(uri)
+	text := docs.Read(uri)
 	lines := strings.Split(strings.ReplaceAll(text, "\r\n", "\n"), "\n")
 	editorLine := params.Position.Line
 	charPos := params.Position.Character
