+++ internal/lsp/lsp.go
@@ -11,6 +11,7 @@ import (
 
 	"github.com/onlyati/quadlet-lsp/internal/commands"
 	"github.com/onlyati/quadlet-lsp/internal/data"
+	"github.com/onlyati/quadlet-lsp/internal/documents"
 	"github.com/onlyati/quadlet-lsp/internal/semantic"
 	"github.com/onlyati/quadlet-lsp/internal/syntax"
 	"github.com/onlyati/quadlet-lsp/internal/utils"
@@ -26,7 +27,7 @@ var (
 	version   = data.ProgramVersion
 	handler   protocol.Handler
 	config    *utils.QuadletConfig
-	documents = utils.NewDocuments()
+	docs      = documents.NewDocuments()
 	commander commands.EditorCommandExecutor
 )
 
@@ -49,10 +50,10 @@ func Start() {
 		// like looking for references.
 		TextDocumentDidOpen: func(ctx *glsp.Context, params *protocol.DidOpenTextDocumentParams) error {
 			uri := string(params.TextDocument.URI)
-			documents.Add(uri, params.TextDocument.Text)
+			docs.Add(uri, params.TextDocument.Text)
 
 			// Check syntax when file is open
-			checker := syntax.NewSyntaxChecker(documents.Read(uri), uri)
+			checker := syntax.NewSyntaxChecker(docs.Read(uri), uri)
 
 			diags := checker.RunAll(config)
 			if len(diags) > 0 {
@@ -71,7 +72,7 @@ func Start() {
 		},
 		TextDocumentDidChange: func(ctx *glsp.Context, params *protocol.DidChangeTextDocumentParams) error {
 			uri := string(params.TextDocument.URI)
-			if text, ok := documents.CheckURI(uri); ok {
+			if text, ok := docs.CheckURI(uri); ok {
 				for _, change := range params.ContentChanges {
 					if change_, ok := change.(protocol.TextDocumentContentChangeEvent); ok {
 						startIndex, endIndex := change_.Range.IndexesIn(text)
@@ -80,11 +81,11 @@ func Start() {
 						text = change_.Text
 					}
 				}
-				documents.Add(uri, text)
+				docs.Add(uri, text)
 			}
 
 			// Check syntax when file is changed
-			checker := syntax.NewSyntaxChecker(documents.Read(uri), uri)
+			checker := syntax.NewSyntaxChecker(docs.Read(uri), uri)
 
 			diags := checker.RunAll(config)
 			if len(diags) > 0 {
@@ -103,7 +104,7 @@ func Start() {
 		},
 		TextDocumentDidClose: func(context *glsp.Context, params *protocol.DidCloseTextDocumentParams) error {
 			uri := string(params.TextDocument.URI)
-			documents.Delete(uri)
+			docs.Delete(uri)
 			return nil
 		},
 
@@ -172,7 +173,7 @@ func initialize(context *glsp.Context, params *protocol.InitializeParams) (any, 
 		context,
 		path.Join(workspaceDir, ".quadletrc.json"),
 		config,
-		&documents,
+		&docs,
 	)
 
 	// Setup server
